fix: 데이터 관리 일반 사용자 palette, margin 완료
